added filter, login page and login controller looks like it even works
