+++ src/main/java/ecommerce/controllers/LoginController.java
@@ -1,0 +1,43 @@
+package ecommerce.controllers;
+
+import ecommerce.exceptions.AuthenticationException;
+import ecommerce.lib.Injector;
+import ecommerce.model.User;
+import ecommerce.security.AuthenticationService;
+import java.io.IOException;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+@WebServlet("/login")
+public class LoginController extends HttpServlet {
+    private static final Injector injector = Injector.getInstance("ecommerce");
+    private final AuthenticationService authenticationService
+            = (AuthenticationService) injector.getInstance(AuthenticationService.class);
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
+            throws ServletException, IOException {
+        req.getRequestDispatcher("/WEB-INF/views/login.jsp").forward(req,resp);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp)
+            throws ServletException, IOException {
+        String login = req.getParameter("login");
+        String password = req.getParameter("password");
+        try {
+            User user = authenticationService.login(login, password);
+            HttpSession httpSession = req.getSession();
+            httpSession.setAttribute("user_id", user.getId());
+        } catch (AuthenticationException e) {
+            req.setAttribute("errorMessage", e.getMessage());
+            req.getRequestDispatcher("/WEB-INF/views/login.jsp").forward(req,resp);
+            return;
+        }
+        resp.sendRedirect(req.getContextPath() + "/");
+    }
+}
