+++ src/main/java/ecommerce/controllers/order/DeleteOrderController.java
@@ -1,7 +1,6 @@
-package ecommerce.controllers;
+package ecommerce.controllers.order;
 
 import ecommerce.lib.Injector;
-import ecommerce.model.Order;
 import ecommerce.service.interfaces.OrderService;
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -10,8 +9,8 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/user/order")
-public class GetOrderDetailsController extends HttpServlet {
+@WebServlet("/order/delete")
+public class DeleteOrderController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("ecommerce");
     private final OrderService orderService
             = (OrderService) injector.getInstance(OrderService.class);
@@ -19,11 +18,8 @@ public class GetOrderDetailsController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        Long orderId = Long.parseLong(req.getParameter("id"));
-        Order order = orderService.get(orderId);
-        req.setAttribute("order", order);
-        Double totalSum = orderService.getTotalSum(order);
-        req.setAttribute("totalSum", totalSum);
-        req.getRequestDispatcher("/WEB-INF/views/order/details.jsp").forward(req, resp);
+        Long orderId = Long.parseLong(req.getParameter("orderId"));
+        orderService.delete(orderId);
+        resp.sendRedirect(req.getContextPath() + "/order/all");
     }
 }
