+++ src/main/java/ecommerce/controllers/order/GetAllOrdersController.java
@@ -1,8 +1,8 @@
-package ecommerce.controllers;
+package ecommerce.controllers.order;
 
 import ecommerce.lib.Injector;
-import ecommerce.model.Product;
-import ecommerce.service.interfaces.ProductService;
+import ecommerce.model.Order;
+import ecommerce.service.interfaces.OrderService;
 import java.io.IOException;
 import java.util.List;
 import javax.servlet.ServletException;
@@ -11,18 +11,17 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/product/all")
-public class GetAllProductsController extends HttpServlet {
+@WebServlet("/admin/order/all")
+public class GetAllOrdersController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("ecommerce");
-    private final ProductService productService
-            = (ProductService) injector.getInstance(ProductService.class);
+    private final OrderService orderService
+            = (OrderService) injector.getInstance(OrderService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        List<Product> allProducts = productService.getAll();
-        req.setAttribute("products", allProducts);
-        req.getRequestDispatcher("/WEB-INF/views/product/all.jsp").forward(req, resp);
+        List<Order> orders = orderService.getAll();
+        req.setAttribute("orders", orders);
+        req.getRequestDispatcher("/WEB-INF/views/order/all.jsp").forward(req, resp);
     }
-
 }
