+++ src/main/java/ecommerce/controllers/order/GetOrderDetailsController.java
@@ -1,18 +1,17 @@
-package ecommerce.controllers;
+package ecommerce.controllers.order;
 
 import ecommerce.lib.Injector;
 import ecommerce.model.Order;
 import ecommerce.service.interfaces.OrderService;
 import java.io.IOException;
-import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/admin/order/all")
-public class GetAllOrdersController extends HttpServlet {
+@WebServlet("/user/order")
+public class GetOrderDetailsController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("ecommerce");
     private final OrderService orderService
             = (OrderService) injector.getInstance(OrderService.class);
@@ -20,8 +19,11 @@ public class GetAllOrdersController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        List<Order> orders = orderService.getAll();
-        req.setAttribute("orders", orders);
-        req.getRequestDispatcher("/WEB-INF/views/order/all.jsp").forward(req, resp);
+        Long orderId = Long.parseLong(req.getParameter("id"));
+        Order order = orderService.get(orderId);
+        req.setAttribute("order", order);
+        Double totalSum = orderService.getTotalSum(order);
+        req.setAttribute("totalSum", totalSum);
+        req.getRequestDispatcher("/WEB-INF/views/order/details.jsp").forward(req, resp);
     }
 }
