+++ src/main/java/ecommerce/controllers/product/AddProductToCartController.java
@@ -1,6 +1,8 @@
-package ecommerce.controllers;
+package ecommerce.controllers.product;
 
 import ecommerce.lib.Injector;
+import ecommerce.model.Product;
+import ecommerce.model.ShoppingCart;
 import ecommerce.service.interfaces.ProductService;
 import ecommerce.service.interfaces.ShoppingCartService;
 import java.io.IOException;
@@ -9,21 +11,23 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/cart/delete")
-public class DeleteProductFromCartController extends HttpServlet {
+@WebServlet("/product/buy")
+public class AddProductToCartController extends HttpServlet {
     private static final Long USER_ID = 1L;
     private static final Injector injector = Injector.getInstance("ecommerce");
-    private final ProductService productService
-            = (ProductService) injector.getInstance(ProductService.class);
     private final ShoppingCartService shoppingCartService
             = (ShoppingCartService) injector.getInstance(ShoppingCartService.class);
+    private final ProductService productService
+            = (ProductService) injector.getInstance(ProductService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws IOException {
-        Long productId = Long.parseLong(req.getParameter("id"));
-        shoppingCartService.deleteProduct(shoppingCartService.getByUserId(USER_ID),
-                productService.get(productId));
-        resp.sendRedirect(req.getContextPath() + "/cart/current");
+        Long productId = Long.valueOf(req.getParameter("id"));
+        ShoppingCart currentCart = shoppingCartService.getByUserId(USER_ID);
+        Product currentProduct = productService.get(productId);
+        shoppingCartService.addProduct(currentCart, currentProduct);
+
+        resp.sendRedirect(req.getContextPath() + "/product/all");
     }
 }
