+++ src/main/java/ecommerce/controllers/product/AdminProductListController.java
@@ -1,0 +1,27 @@
+package ecommerce.controllers.product;
+
+import ecommerce.lib.Injector;
+import ecommerce.model.Product;
+import ecommerce.service.interfaces.ProductService;
+import java.io.IOException;
+import java.util.List;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@WebServlet("/admin/product/all")
+public class AdminProductListController extends HttpServlet {
+    private static final Injector injector = Injector.getInstance("ecommerce");
+    private final ProductService productService
+            = (ProductService) injector.getInstance(ProductService.class);
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
+            throws ServletException, IOException {
+        List<Product> products = productService.getAll();
+        req.setAttribute("products", products);
+        req.getRequestDispatcher("/WEB-INF/views/product/list.jsp").forward(req, resp);
+    }
+}
