+++ src/main/java/ecommerce/controllers/user/GetAllUsersController.java
@@ -1,25 +1,27 @@
-package ecommerce.controllers;
+package ecommerce.controllers.user;
 
 import ecommerce.lib.Injector;
-import ecommerce.service.interfaces.OrderService;
+import ecommerce.model.User;
+import ecommerce.service.interfaces.UserService;
 import java.io.IOException;
+import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/order/delete")
-public class DeleteOrderController extends HttpServlet {
+@WebServlet("/user/all")
+public class GetAllUsersController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("ecommerce");
-    private final OrderService orderService
-            = (OrderService) injector.getInstance(OrderService.class);
+    private final UserService userService = (UserService) injector.getInstance(UserService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        Long orderId = Long.parseLong(req.getParameter("orderId"));
-        orderService.delete(orderId);
-        resp.sendRedirect(req.getContextPath() + "/order/all");
+        List<User> allUsers = userService.getAll();
+
+        req.setAttribute("users", allUsers);
+        req.getRequestDispatcher("/WEB-INF/views/user/all.jsp").forward(req, resp);
     }
 }
