+++ src/main/java/ecommerce/dao/interfaces/UserDao.java
@@ -2,6 +2,8 @@ package ecommerce.dao.interfaces;
 
 import ecommerce.dao.GenericDao;
 import ecommerce.model.User;
+import java.util.Optional;
 
 public interface UserDao extends GenericDao<User, Long> {
+    Optional<User> findByLogin(String login);
 }
