+++ src/main/java/ecommerce/exceptions/AuthenticationException.java
@@ -1,0 +1,7 @@
+package ecommerce.exceptions;
+
+public class AuthenticationException extends Exception {
+    public AuthenticationException(String message) {
+        super(message);
+    }
+}
