+++ src/main/java/ecommerce/security/AuthenticationService.java
@@ -1,0 +1,9 @@
+package ecommerce.security;
+
+import ecommerce.exceptions.AuthenticationException;
+import ecommerce.model.User;
+
+public interface AuthenticationService {
+    User login(String login, String password)
+            throws AuthenticationException;
+}
