+++ src/main/java/ecommerce/service/impl/UserServiceImpl.java
@@ -6,6 +6,7 @@ import ecommerce.lib.Service;
 import ecommerce.model.User;
 import ecommerce.service.interfaces.UserService;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -36,5 +37,10 @@ public class UserServiceImpl implements UserService {
     @Override
     public boolean delete(Long id) {
         return userDao.delete(id);
+    }
+
+    @Override
+    public Optional<User> findByLogin(String login) {
+        return userDao.findByLogin(login);
     }
 }
