+++ src/main/java/ecommerce/service/interfaces/UserService.java
@@ -2,6 +2,8 @@ package ecommerce.service.interfaces;
 
 import ecommerce.model.User;
 import ecommerce.service.GenericService;
+import java.util.Optional;
 
 public interface UserService extends GenericService<User, Long> {
+    Optional<User> findByLogin(String login);
 }
